Fix checks, auto format most of the code
IJ would not auto format the whole code base for some reason

+++ stroom-util/src/main/java/stroom/util/io/ByteSizeUnit.java
@@ -86,7 +86,7 @@ public enum ByteSizeUnit {
      * @return The number of bytes in this byte size unit
      */
     public long longBytes() {
-        return (long) bytes;
+        return bytes;
     }
 
     /**
